requirements.txt
Django>=4.1,<4.4
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<3.0
drf-spectacular>=0.15.1,<0.24
Pillow>=8.5,<9.5